[seller] refactor(173): calculation 관련 hook을 Public API 규칙으로 수정
